+++ HACKING
@@ -110,6 +110,8 @@ infrastructure in place for it from the start.  Use it like this:
 
 ## Dependencies
 
+* *Augeas* - Round trip configuration editing utilities
+
 * *Bootstrap Form* - Render Django forms for Twitter Bootstrap
 
 * *CherryPy3* - WSGI web server since Django does not have proper web server
