+++ INSTALL
@@ -5,10 +5,10 @@
     On a Debian based system, run:
 
     $ sudo apt-get install libjs-jquery libjs-modernizr libjs-bootstrap \
-    make pandoc python3 python3-bootstrapform python3-cherrypy3 \
-    python3-coverage python3-django python3-django-stronghold python3-gi \
-    python3-setuptools python3-yaml gir1.2-glib-2.0 gir1.2-networkmanager-1.0 \
-    gir1.2-packagekitglib-1.0
+    make pandoc python3 python3-augeas python3-bootstrapform \
+    python3-cherrypy3 python3-coverage python3-django \
+    python3-django-stronghold python3-gi python3-setuptools python3-yaml \
+    gir1.2-glib-2.0 gir1.2-networkmanager-1.0 gir1.2-packagekitglib-1.0
 
 2. Install Plinth:
