+++ plinth/modules/pagekite/utils.py
@@ -18,6 +18,7 @@
 from gettext import gettext as _
 import json
 import logging
+import os
 
 from plinth import actions
 
@@ -29,6 +30,10 @@ BACKEND_HOST = 'localhost'
 KITE_NAME = '@kitename'
 KITE_SECRET = '@kitesecret'
 
+# Augeas base path for Pagekite configuration files
+CONF_PATH = '/files/etc/pagekite.d'
+
+# Parameters that get stored in configuration service_on entries
 SERVICE_PARAMS = ['protocol', 'kitename', 'backend_host', 'backend_port',
                   'secret']
 
@@ -72,7 +77,7 @@ PREDEFINED_SERVICES = {
 
 
 def get_kite_details():
-    output = _run(['get-kite'])
+    output = run(['get-kite'])
     kite_details = output.split()
     return {'kite_name': kite_details[0],
             'kite_secret': kite_details[1]}
@@ -86,14 +91,14 @@ def get_pagekite_config():
 
     # PageKite service enabled/disabled
     # This assumes that if pagekite is running it's also enabled as a service
-    output = _run(['is-running'])
+    output = run(['is-running'])
     status['enabled'] = (output.split()[0] == 'yes')
 
     # PageKite kite details
     status.update(get_kite_details())
 
     # PageKite frontend server
-    server = _run(['get-frontend'])
+    server = run(['get-frontend'])
     status['server'] = server.replace('\n', '')
 
     return status
@@ -110,7 +115,7 @@ def get_pagekite_services():
     # set all predefined services to 'disabled' by default
     [predefined.update({proto: False}) for proto in PREDEFINED_SERVICES.keys()]
     # now, search for the enabled ones
-    for serviceline in _run(['get-services']).split('\n'):
+    for serviceline in run(['get-services']).split('\n'):
         if not serviceline:  # skip empty lines
             continue
 
@@ -137,7 +142,7 @@ def prepare_service_for_display(service):
     return service
 
 
-def _run(arguments, superuser=True):
+def run(arguments, superuser=True):
     """Run a given command and raise exception if there was an error"""
     command = 'pagekite'
 
@@ -145,3 +150,78 @@ def _run(arguments, superuser=True):
         return actions.superuser_run(command, arguments)
     else:
         return actions.run(command, arguments)
+
+
+def convert_service_to_string(service):
+    """ Convert service dict into a ":"-separated parameter string
+
+    >>> convert_service_to_string({'kitename': '@kitename', \
+'backend_host': 'localhost', 'secret': '@kitesecret', \
+'protocol': 'https/443', 'backend_port': '443'})
+    'https/443:@kitename:localhost:443:@kitesecret'
+    """
+    try:
+        service_string = ":".join([service[param] for param in SERVICE_PARAMS])
+    except KeyError:
+        raise ValueError("Could not parse params: %s " % service)
+    return service_string
+
+
+def load_service(json_service):
+    """ create a service out of json command-line argument
+
+    1) parse json
+    2) only use the parameters that we need (SERVICE_PARAMS)
+    3) convert unicode to strings
+    """
+    service = json.loads(json_service)
+    return dict((str(key), str(service[key])) for key in SERVICE_PARAMS)
+
+
+def get_augeas_servicefile_path(protocol):
+    """Get the augeas path where a service for a protocol should be stored
+
+    TODO: Once we use python3 switch from doctests to unittests
+
+    >>> get_augeas_servicefile_path('http')
+    '/files/etc/pagekite.d/80_http.rc/service_on'
+
+    >>> get_augeas_servicefile_path('https')
+    '/files/etc/pagekite.d/443_https.rc/service_on'
+
+    >>> get_augeas_servicefile_path('http/80')
+    '/files/etc/pagekite.d/80_http.rc/service_on'
+
+    >>> get_augeas_servicefile_path('http/8080')
+    '/files/etc/pagekite.d/8080_http.rc/service_on'
+
+    >>> get_augeas_servicefile_path('raw/22')
+    '/files/etc/pagekite.d/22_raw.rc/service_on'
+
+    >>> get_augeas_servicefile_path('xmpp')
+    Traceback (most recent call last):
+        ...
+    ValueError: Unsupported protocol: xmpp
+
+    """
+    if not protocol.startswith(("http", "https", "raw")):
+        raise ValueError('Unsupported protocol: %s' % protocol)
+
+    try:
+        _protocol, port = protocol.split('/')
+    except ValueError:
+        if protocol == 'http':
+            relpath = '80_http.rc'
+        elif protocol == 'https':
+            relpath = '443_https.rc'
+        else:
+            raise ValueError('Unsupported protocol: %s' % protocol)
+    else:
+        relpath = '%s_%s.rc' % (port, _protocol)
+
+    return os.path.join(CONF_PATH, relpath, 'service_on')
+
+
+if __name__ == "__main__":
+    import doctest
+    doctest.testmod()
