+++ requirements.txt
@@ -2,4 +2,5 @@ cherrypy >= 3.0
 coverage >= 3.7
 django >= 1.7.0
 django-stronghold
+python-augeas
 pyyaml
